Move all converters into converters.py
